Aula 19 configuração ambinete

+++ scripts/set_my_ip.py
@@ -20,16 +20,22 @@ def get_current_ip():
 # Função para atualizar o arquivo .env com o novo IP
 def update_env_file(ip):
     try:
-        with open('.env', 'r') as file:
-            lines = file.readlines()
-        
-        with open('.env', 'w') as file:
-            for line in lines:
-                if line.startswith('MY_IP='):
-                    continue
-                else:
-                    file.write(line)
-            file.write(f'MY_IP={ip}\n')
+        # Verifica se o arquivo .env existe
+        if not os.path.isfile('.env'):
+            # Cria o arquivo .env se ele não existir
+            with open('.env', 'w') as new_file:
+                new_file.write(f'MY_IP={ip}\n')
+        else:
+            with open('.env', 'r') as file:
+                lines = file.readlines()
+            
+            with open('.env', 'w') as file:
+                for line in lines:
+                    if line.startswith('MY_IP='):
+                        continue
+                    else:
+                        file.write(line)
+                file.write(f'MY_IP={ip}\n')
 
         print(f'IP atualizado no arquivo .env: MY_IP={ip}')
     except Exception as e:
